pyproject.toml
[project]
name = "chainlink-chromion-hackatho-back"
version = "0.1.0"
description = ""
authors = [
    {name = "Oscar Riojas",email = "oscarriojas@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "fastapi (>=0.115.12,<0.116.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
